add assert for invalid fields:
