+++ packages/algorand-js/src/lib/constants.ts
@@ -9,6 +9,58 @@ const zeroAddress = new Uint8Array(32);
 const zeroUint64 = BigInt('0');
 const zeroByte = new Uint8Array(0);
 
+// https://developer.algorand.org/docs/reference/teal/opcodes/#txn
+export const TxnFields: {[key: string]: string} = {
+  Sender: 'snd',
+  Fee: 'fee',
+  FirstValid: 'fv',
+  FirstValidTime: '',
+  LastValid: 'lv',
+  Note: 'note',
+  Lease: 'lx',
+  Receiver: 'rcv',
+  Amount: 'amt',
+  CloseRemainderTo: 'close',
+  VotePK: 'votekey',
+  SelectionPK: 'selkey',
+  VoteFirst: 'votefst',
+  VoteLast: 'votelst',
+  VoteKeyDilution: 'votekd',
+  Type: 'type',
+  TypeEnum: '',
+  XferAsset: 'xaid',
+  AssetAmount: 'aamt',
+  AssetSender: 'asnd',
+  AssetReceiver: 'arcv',
+  AssetCloseTo: 'aclose',
+  GroupIndex: '',
+  TxID: '',
+  ApplicationID: 'apid',
+  OnCompletion: 'apan',
+  ApplicationArgs: 'apaa',
+  NumAppArgs: '',
+  Accounts: 'apat',
+  NumAccounts: '',
+  ApprovalProgram: 'apap',
+  ClearStateProgram: 'apsu',
+  RekeyTo: 'rekey',
+  ConfigAsset: 'caid',
+  ConfigAssetTotal: 't',
+  ConfigAssetDecimals: 'dc',
+  ConfigAssetDefaultFrozen: 'df',
+  ConfigAssetUnitName: 'un',
+  ConfigAssetName: 'an',
+  ConfigAssetURL: 'au',
+  ConfigAssetMetadataHash: 'am',
+  ConfigAssetManager: 'm',
+  ConfigAssetReserve: 'r',
+  ConfigAssetFreeze: 'f',
+  ConfigAssetClawback: 'c',
+  FreezeAsset: 'faid',
+  FreezeAssetAccount: 'fadd',
+  FreezeAssetFrozen: 'afrz'
+};
+
 export const TxFieldDefaults: {[key: string]: any} = {
   Sender: zeroAddress,
   Fee: zeroUint64,
