+++ packages/algorand-js/src/types.ts
@@ -4,6 +4,7 @@ import {
   Add, Addr, Arg, Byte, Bytec, Bytecblock, Div, Int, Len, Mul, Pragma,
   Sub
 } from "./interpreter/opcode-list";
+import { TxnFields } from "./lib/constants";
 import type { IStack } from "./lib/stack";
 
 export type Operator = Len | Add | Sub |
@@ -17,58 +18,6 @@ export type TEALStack = IStack<bigint | Uint8Array>;
 export interface Txn extends TxnEncodedObj {
   txID: string
 }
-
-// https://developer.algorand.org/docs/reference/teal/opcodes/#txn
-export const TxnFields: {[key: string]: string} = {
-  Sender: 'snd',
-  Fee: 'fee',
-  FirstValid: 'fv',
-  FirstValidTime: '',
-  LastValid: 'lv',
-  Note: 'note',
-  Lease: 'lx',
-  Receiver: 'rcv',
-  Amount: 'amt',
-  CloseRemainderTo: 'close',
-  VotePK: 'votekey',
-  SelectionPK: 'selkey',
-  VoteFirst: 'votefst',
-  VoteLast: 'votelst',
-  VoteKeyDilution: 'votekd',
-  Type: 'type',
-  TypeEnum: '',
-  XferAsset: 'xaid',
-  AssetAmount: 'aamt',
-  AssetSender: 'asnd',
-  AssetReceiver: 'arcv',
-  AssetCloseTo: 'aclose',
-  GroupIndex: '',
-  TxID: '',
-  ApplicationID: 'apid',
-  OnCompletion: 'apan',
-  ApplicationArgs: 'apaa',
-  NumAppArgs: '',
-  Accounts: 'apat',
-  NumAccounts: '',
-  ApprovalProgram: 'apap',
-  ClearStateProgram: 'apsu',
-  RekeyTo: 'rekey',
-  ConfigAsset: 'caid',
-  ConfigAssetTotal: 't',
-  ConfigAssetDecimals: 'dc',
-  ConfigAssetDefaultFrozen: 'df',
-  ConfigAssetUnitName: 'un',
-  ConfigAssetName: 'an',
-  ConfigAssetURL: 'au',
-  ConfigAssetMetadataHash: 'am',
-  ConfigAssetManager: 'm',
-  ConfigAssetReserve: 'r',
-  ConfigAssetFreeze: 'f',
-  ConfigAssetClawback: 'c',
-  FreezeAsset: 'faid',
-  FreezeAssetAccount: 'fadd',
-  FreezeAssetFrozen: 'afrz'
-};
 
 export type TxField = keyof typeof TxnFields;
